requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.6
    # via requests
mlb-statsapi==1.7.1
    # via -r requirements.in
requests==2.31.0
    # via mlb-statsapi
urllib3==2.1.0
    # via requests
